fix(utils): fix bugs and some warnings
